[FIX] Hard-coded channels used for average referencing
When computing the average reference, the channels used for this purpose were hard-coded in `prep_avgref()`, specifically for a 128-channel net from EGI. To enhance flexibility and accommodate other net configurations, a new variable was added to the configuration file. This change allows the toolbox to adapt to different electrode setups.